adds viewpager refactor code for adopting layouts for viewpager
