+++ app/src/main/java/com/hellcorp/selfdictation/di/viewModelModule.kt
@@ -1,11 +1,13 @@
 package com.hellcorp.selfdictation.di
 
+import com.hellcorp.selfdictation.ui.card.viewmodel.CardViewModel
 import com.hellcorp.selfdictation.ui.main.viewmodels.MainViewModel
-import com.hellcorp.selfdictation.ui.main.viewmodels.UsersSetViewmodel
+import com.hellcorp.selfdictation.ui.usersetlist.UsersSetViewmodel
 import org.koin.androidx.viewmodel.dsl.viewModelOf
 import org.koin.dsl.module
 
 val viewModelModule = module {
     viewModelOf(::MainViewModel)
     viewModelOf(::UsersSetViewmodel)
+    viewModelOf(::CardViewModel)
 }
