+++ app/src/main/res/layout/fragment_users_set.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.main.fragments.UsersSetFragment">
+    tools:context=".ui.usersetlist.UsersSetFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:layout_width="match_parent"
