+++ app/src/main/res/values/colors.xml
@@ -3,6 +3,7 @@
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
     <color name="blue">#2700B8D4</color>
+    <color name="gray">#575757</color>
     <color name="light_green">#805FE10D</color>
     <color name="transparent">@android:color/transparent</color>
     <color name="main_background">#80BCBCBC</color>
